refactor: 자동차 묶음을 일급 컬렉션 Cars로 분리
